Refactored hivemall.mf package
