+++ src/main/java/hivemall/mf/FactorizedModel.java
@@ -16,9 +16,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package hivemall.io;
+package hivemall.mf;
 
-import hivemall.common.RatingInitilizer;
 import hivemall.utils.collections.IntOpenHashMap;
 import hivemall.utils.math.MathUtils;
