+++ src/main/java/hivemall/mf/RatingInitilizer.java
@@ -16,9 +16,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package hivemall.common;
+package hivemall.mf;
 
-import hivemall.io.Rating;
 
 public interface RatingInitilizer {
